Create a gt table object with gsm style

@description

Usage

gsm_gt(data, ...)

Arguments

data

A `gt_tbl` object.

...

Additional arguments passed to [gt::gt()].

Value

An object of class `gt_tbl`.

Details

This is a wrapper to ensure that the user has gt installed and to apply standardized styles.